+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed 
+- exposed `vault_parameters` and `llm_request` fields to redact api
+
 ## 5.2.0 - 2024-12-18
 
 ### Added